Dedup optional param normalization into shared helper
Unify the Feather optional parameter sentinel handling from src/plugin/src/printer/print.js and src/plugin/src/comments/line-comment-formatting.js by routing both call sites through src/plugin/src/comments/optional-param-normalization.js.
