+++ src/plugin/src/comments/index.js
@@ -44,3 +44,4 @@ export {
     resolveDocCommentUpdateService
 } from "./doc-comment-manager.js";
 export { getCommentValue } from "@prettier-plugin-gml/shared/ast/comments.js";
+export { normalizeOptionalParamToken } from "./optional-param-normalization.js";
