+++ src/plugin/src/printer/print.js
@@ -34,6 +34,7 @@ import {
     formatLineComment,
     normalizeDocCommentTypeAnnotations
 } from "../comments/line-comment-formatting.js";
+import { normalizeOptionalParamToken } from "../comments/optional-param-normalization.js";
 import { resolveLineCommentOptions } from "../options/line-comment-options.js";
 import { TRAILING_COMMA } from "../options/trailing-comma-option.js";
 import { DEFAULT_DOC_COMMENT_MAX_WRAP_WIDTH } from "./doc-comment-wrap-width.js";
@@ -6292,7 +6293,7 @@ function parseDocCommentMetadata(line) {
             name = paramMatch ? paramMatch[1] : null;
         }
         if (typeof name === "string") {
-            name = normalizeOptionalParamNameToken(name);
+            name = normalizeOptionalParamToken(name);
         }
 
         return {
@@ -6303,43 +6304,6 @@ function parseDocCommentMetadata(line) {
     }
 
     return { tag, name: remainder };
-}
-
-function normalizeOptionalParamNameToken(name) {
-    if (typeof name !== "string") {
-        return name;
-    }
-
-    const trimmed = name.trim();
-
-    if (/^\[[^\]]+\]$/.test(trimmed)) {
-        return trimmed;
-    }
-
-    let stripped = trimmed;
-    let hadSentinel = false;
-
-    while (stripped.startsWith("*")) {
-        stripped = stripped.slice(1);
-        hadSentinel = true;
-    }
-
-    while (stripped.endsWith("*")) {
-        stripped = stripped.slice(0, -1);
-        hadSentinel = true;
-    }
-
-    if (!hadSentinel) {
-        return trimmed;
-    }
-
-    const normalized = stripped.trim();
-
-    if (normalized.length === 0) {
-        return stripped.replaceAll("*", "");
-    }
-
-    return `[${normalized}]`;
 }
 
 function getSourceTextForNode(node, options) {
@@ -7305,7 +7269,7 @@ function normalizeDocMetadataName(name) {
         return name;
     }
 
-    const optionalNormalized = normalizeOptionalParamNameToken(name);
+    const optionalNormalized = normalizeOptionalParamToken(name);
     if (typeof optionalNormalized === "string") {
         if (/^\[[^\]]+\]$/.test(optionalNormalized)) {
             return optionalNormalized;
